pyproject.toml
[build-system]
requires = ["hatchling>=1.26"]
build-backend = "hatchling.build"

[project]
name = "harbor"
version = "0.2.0"
description = "Stateful agent-graph framework with deterministic governance via Fathom"
readme = "README.md"
requires-python = ">=3.13,<3.14"
license = "Apache-2.0"
license-files = ["LICENSE"]
authors = [{ name = "Kraken Networks", email = "info@krakennetworks.com" }]
keywords = ["agents", "graph", "ir", "fathom", "governance", "ai", "policy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.9",
    "pluggy>=1.5",
    "fathom-rules>=0.3.1",
    "jsonschema>=4.21",
    "pyyaml>=6.0",
    "structlog>=24.1",
    "anyio>=4.3",
    "uuid-utils>=0.8",
    "dspy>=3.0.4,<3.4",
    "mcp>=1.0",
    "aiosqlite>=0.20",
    "asyncpg>=0.29",
    "rfc8785>=0.1.4,<0.2",
    "orjson>=3.10",
    "typer>=0.12",
    "jsonpatch>=1.33",
    "cryptography>=47.0.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "httpx>=0.27",
    "cronsim>=2.6",
    "pyjwt[crypto]>=2.8",
    "nautilus-rkm==0.1.2",
    "blake3>=0.4",
    "argon2-cffi>=23.1",
]

[tool.uv.sources]
# nautilus-rkm is not on PyPI; resolve from the sibling worktree path.
# Path is relative to this pyproject.toml (in .worktrees/harbor-engine/).
nautilus-rkm = { path = "../../nautilus", editable = false }

[project.optional-dependencies]
ml = [
    "scikit-learn",
    "xgboost>=2.1",
    "onnxruntime>=1.18",
    "skops>=0.10",
    "joblib",
]
# Note on `kuzu==0.11.3`: Kuzu's GitHub repo was archived 2025-10-10 after
# Apple acquired Kuzu Inc. (Oct 2025). Active community fork lives at
# `predictable-labs/ryugraph`. Harbor abstracts Kuzu behind GraphStore so
# swapping to RyuGraph is a one-file change. Pin tight until the fork story
# stabilizes. (FR-11, AC-12.1)
stores = [
    "lancedb>=0.30.2,<0.40",
    "kuzu==0.11.3",
    "pyarrow>=15",
]
skills-rag = [
    "sentence-transformers>=5.0,<6",
    "huggingface_hub>=0.25",
]

[project.urls]
Homepage = "https://harbor.krakn.ai"
Repository = "https://github.com/KrakenNet/harbor"
Documentation = "https://harbor.krakn.ai"
Changelog = "https://github.com/KrakenNet/harbor/blob/main/CHANGELOG.md"
Issues = "https://github.com/KrakenNet/harbor/issues"

[project.scripts]
harbor = "harbor.cli:main"

# Public extension groups — plugins register here
[project.entry-points."harbor"]
# (reserved — Harbor itself does not register here; downstream plugins use it
#  for the harbor_plugin manifest entry-point)

# Note: harbor.{tools,skills,stores,packs} have NO core entries — they are
# entirely populated by external plugin distributions.

# Built-in trigger plugins (design §6.3, FR-3/4/5). Cron + Webhook land in
# tasks 2.10 / 2.11; manual is the simplest baseline (CLI + HTTP convergence).
[project.entry-points."harbor.triggers"]
manual = "harbor.triggers.manual:ManualTrigger"
cron = "harbor.triggers.cron:CronTrigger"
webhook = "harbor.triggers.webhook:WebhookTrigger"

# Built-in Nautilus tool plugin (design §8.2, FR-45). Registry-discoverable
# `@tool`-decorated coroutine that calls the lifespan-singleton Broker.
[project.entry-points."harbor.tools"]
broker_request = "harbor.tools.nautilus.broker_request:broker_request"

[dependency-groups]
dev = [
    "ruff>=0.6",
    "pyright>=1.1.380",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "hypothesis>=6.0",
    "pre-commit>=3.7",
    "codespell>=2.2",
    "vcrpy>=6.0",
    "pytest-recording>=0.13",
    "testcontainers[postgresql]>=4.0",
    "testcontainers[neo4j]>=4.0",
    "trustme>=1.2",
    "time-machine>=3.2.0",
    "pytest-benchmark>=5.2.3",
]
docs = [
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.24",
    "mkdocs-redirects>=1.2",
    "mike>=2.1",
]
release = [
    "build>=1.2",
    "twine>=5.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/harbor"]

[tool.hatch.build.targets.sdist]
include = ["src/harbor", "tests", "README.md", "LICENSE", "CHANGELOG.md"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH", "ASYNC", "RUF"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]  # asserts allowed in tests

[tool.pyright]
strict = ["src/harbor", "tests"]
include = ["src/harbor", "tests"]
pythonVersion = "3.12"
venvPath = "."
venv = ".venv"
reportMissingTypeStubs = "warning"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--strict-markers --cov=harbor --cov-branch -ra"
markers = [
    "unit: fast unit tests (no external deps)",
    "integration: requires Fathom CLIPS / external services",
    "slow: long-running tests",
    "knowledge: harbor-knowledge subsystem tests (stores, skills, retrieval, memory)",
    "serve: harbor-serve subsystem tests (umbrella for serve-related markers)",
    "websocket: WebSocket / streaming tests",
    "api: FastAPI HTTP API tests",
    "trigger: trigger plugin tests (manual/cron/webhook)",
    "scheduler: scheduler / queue tests",
]

[tool.coverage.run]
source = ["harbor"]
branch = true

[tool.coverage.report]
show_missing = true
skip_empty = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
    "@overload",
]
